update dl all button

+++ code/gridfield/GridFieldExportAllButton.php
@@ -5,6 +5,10 @@
  */
 class GridFieldExportAllButton extends GridFieldExportButton
 {
+    /**
+     * @var string
+     */
+    protected $csvSeparator = ";";
 
     /**
      * Place the export button in a <p> tag below the field
@@ -12,14 +16,40 @@ class GridFieldExportAllButton extends GridFieldExportButton
     public function getHTMLFragments($gridField)
     {
         $button = new GridField_FormAction(
-            $gridField, 'export',
-            _t('GridFieldExportAllButton.LABEL', 'Export all to CSV'), 'export',
-            null
+            $gridField, 'export_all',
+            _t('GridFieldExportAllButton.LABEL', 'Export all to CSV'),
+            'export_all', null
         );
         $button->setAttribute('data-icon', 'download-csv');
         $button->addExtraClass('no-ajax');
         return array(
             $this->targetFragment => '<p class="grid-csv-button">'.$button->Field().'</p>',
+        );
+    }
+
+    /**
+     * export is an action button
+     */
+    public function getActions($gridField)
+    {
+        return array('export_all');
+    }
+
+    public function handleAction(GridField $gridField, $actionName, $arguments,
+                                 $data)
+    {
+        if ($actionName == 'export_all') {
+            return $this->handleExport($gridField);
+        }
+    }
+
+    /**
+     * it is also a URL
+     */
+    public function getURLHandlers($gridField)
+    {
+        return array(
+            'export_all' => 'handleExport',
         );
     }
 
@@ -31,8 +61,14 @@ class GridFieldExportAllButton extends GridFieldExportButton
      */
     public function generateExportFileData($gridField)
     {
-        $separator  = $this->csvSeparator;
-        $csvColumns = ($this->exportColumns) ? $this->exportColumns : singleton($gridField->getModelClass())->summaryFields();
+        $separator = $this->csvSeparator;
+        $singl     = singleton($gridField->getModelClass());
+        if ($singl->hasMethod('exportedFields')) {
+            $fallbackColumns = $singl->exportedFields();
+        } else {
+            $fallbackColumns = $singl->summaryFields();
+        }
+        $csvColumns = ($this->exportColumns) ? $this->exportColumns : $fallbackColumns;
         $fileData   = '';
         $columnData = array();
         $fieldItems = new ArrayList();
@@ -54,27 +90,39 @@ class GridFieldExportAllButton extends GridFieldExportButton
         $items = $gridField->getList();
 
         foreach ($items as $item) {
-            $columnData = array();
-            foreach ($csvColumns as $columnSource => $columnHeader) {
-                if (!is_string($columnHeader) && is_callable($columnHeader)) {
-                    if ($item->hasMethod($columnSource)) {
-                        $relObj = $item->{$columnSource}();
+            if (!$item->hasMethod('canView') || $item->canView()) {
+                $columnData = array();
+
+                foreach ($csvColumns as $columnSource => $columnHeader) {
+                    if (!is_string($columnHeader) && is_callable($columnHeader)) {
+                        if ($item->hasMethod($columnSource)) {
+                            $relObj = $item->{$columnSource}();
+                        } else {
+                            $relObj = $item->relObject($columnSource);
+                        }
+
+                        $value = $columnHeader($relObj);
                     } else {
-                        $relObj = $item->relObject($columnSource);
+                        $value = $gridField->getDataFieldValue($item,
+                            $columnSource);
+
+                        if (!$value) {
+                            $value = $gridField->getDataFieldValue($item,
+                                $columnHeader);
+                        }
                     }
 
-                    $value = $columnHeader($relObj);
-                } else {
-                    $value = $gridField->getDataFieldValue($item, $columnSource);
+                    $value        = str_replace(array("\r", "\n"), "\n", $value);
+                    $columnData[] = '"'.str_replace('"', '""', utf8_decode($value)).'"';
                 }
 
-                $value        = str_replace(array("\r", "\n"), "\n", $value);
-                $columnData[] = '"'.str_replace('"', '\"', $value).'"';
+                $fileData .= implode($separator, $columnData);
+                $fileData .= "\n";
             }
-            $fileData .= implode($separator, $columnData);
-            $fileData .= "\n";
 
-            $item->destroy();
+            if ($item->hasMethod('destroy')) {
+                $item->destroy();
+            }
         }
 
         return $fileData;
